refactor: route hosted_vllm Anthropic passthrough through ProviderConfigManager
- Move _should_skip_anthropic_translation to hosted_vllm/messages/transformation.py
- Register HostedVLLMAnthropicMessagesConfig in ProviderConfigManager instead of
  hard-coding a provider branch in the Anthropic messages handler
- Pass litellm_params to get_provider_anthropic_messages_config so the flag check
  can happen inside the abstraction used by all other providers
- Pin ProviderConfigManager mock in test_flag_disabled_uses_translation so the test
  is resilient to future hosted_vllm config registrations
